Add flags to flake8 and black for pre-commit
Add flags to flake8 and black for pre-commit based on existing flags in
GitHub Actions workflows. This includes selecting for specific flake8
errors and adjusting the max line length. These existing configs are
extremely limited, so they should be broadened to more errors and/or a
more widely accepted max line length.

Also ran black on the entire codebase to satisfy the GitHub workflows.