Complete TODOs for combo machines and status_toggle = 1
Replace mock JSON test data for Sutherland East with new mock JSON test
data for Holland. Because this data was collected close to the start of
the fall semester, when the laundry machines are actually being used, it
contains instances of status_toggle = 1, which clears up the TODO of
what that status value represented. The meaning of this value has now
been documented.

Rewrite the logic for parsing JSON objects for laundry machines. The
logic now accurately parses combo machines and adds support for double
machines, which are two washers or two dryers in a single machine. This
resolves the two remaining TODOs about combo machines.