Convert gym unit test mock from Python str to HTML
Convert the mock for gym unit tests from a Python string to an HTML
file. This makes the mocking of gym unit tests more consistent with how
unit tests for other modules are mocked. This also avoids unnecessarily
mocking entire module functions, as only the HTTP requests made by those
functions need to be mocked. Since the mock is only supposed to contain
an HTML HTTP response, it makes more sense to keep it as HTML.